発行された年をクエリで受け取るように変更とSOLの型に合わせてLongからint型に変更

+++ lecture09/mybatis/src/main/java/com/example/mybatis/mapper/MovieMapper.java
@@ -21,5 +21,5 @@ public interface MovieMapper {
      * @return 指定された年のMovieをリストで返却する
      */
     @Select("SELECT name FROM movies WHERE published_year=#{publishedYear}")
-    List<String> findMoviesByPublishedYear(long publishedYear);
+    List<String> findMoviesByPublishedYear(int publishedYear);
 }
